added support for alternate db providers
basic entity framework contexts with model builder customizations are now defined for MS-SQL and PG-SQL

+++ src/contract/IUser.cs
@@ -5,7 +5,7 @@ namespace Toucan.Contract
     {
         string Email { get; }
         bool Enabled { get; set; }
-        int UserId { get; }
+        long UserId { get; }
         string Username { get; }
     }
 }
